# openapi.yaml
openapi: 3.1.0
info:
  title: "MAFK Notion API"
  version: "4.0"
servers:
  - url: "https://mafk-notion-api.onrender.com"
paths:
...
  /projects/{id}:
    patch:
      summary: "Update project metadata"
      operationId: updateProject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                Status:
                  type: string
                Deadline:
                  type: string
                  format: date
                Description:
                  type: string
                Owner:
                  type: string
                Client:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          description: "Project updated"
...